+++ lib/activerecord-bitemporal/bitemporal.rb
@@ -227,16 +227,45 @@ module ActiveRecord
         scope :without_deleted, -> {
           with_bitemporal_option(within_deleted: false)
         }
+        scope :bitemporal_for, -> (id) {
+          where(bitemporal_id: id)
+        }
+        scope :valid_in, -> (from: nil, to: nil) {
+          table_name = klass.table_name
+          column_valid_to = "#{table_name}.valid_to"
+          column_valid_from = "#{table_name}.valid_from"
+          ignore_valid_datetime
+            .tap { |relation| break relation.where("? <= #{column_valid_to}", from.in_time_zone.to_datetime) if from }
+            .tap { |relation| break relation.where("#{column_valid_from} <= ?", to.in_time_zone.to_datetime) if to }
+        }
+        scope :valid_allin, -> (from: nil, to: nil) {
+          table_name = klass.table_name
+          column_valid_to = "#{table_name}.valid_to"
+          column_valid_from = "#{table_name}.valid_from"
+          ignore_valid_datetime
+            .tap { |relation| break relation.where("? <= #{column_valid_from}", from.in_time_zone.to_datetime) if from }
+            .tap { |relation| break relation.where("#{column_valid_to} <= ?", to.in_time_zone.to_datetime) if to }
+        }
+      end
+
+      module Extension
+        extend ActiveSupport::Concern
+
+        included do
+          scope :bitemporal_histories, -> (*ids) {
+            ignore_valid_datetime.bitemporal_for(*ids)
+          }
+          def self.bitemporal_most_future(id)
+            bitemporal_histories(id).order(valid_from: :asc).last
+          end
+          def self.bitemporal_most_past(id)
+            bitemporal_histories(id).order(valid_from: :asc).first
+          end
+        end
       end
 
       module Experimental
         extend ActiveSupport::Concern
-
-        included do
-          scope :valid_in, -> (from:, to:) {
-            ignore_valid_datetime.where("valid_from <= ?", to).where("? < valid_to", from)
-          }
-        end
       end
     end
 
@@ -435,14 +464,14 @@ module ActiveRecord
           #   一番近い未来の履歴レコードを参照して更新する
           # という仕様があるため、それを考慮して valid_to を設定する
           if (record.valid_datetime && (record.valid_from..record.valid_to).include?(record.valid_datetime)) == false
-            finder_class.where(bitemporal_id: record.bitemporal_id).where('valid_from > ?', target_datetime).ignore_valid_datetime.order(valid_from: :asc).first.valid_from
+            finder_class.where(bitemporal_id: record.bitemporal_id).where('? < valid_from', target_datetime).ignore_valid_datetime.order(valid_from: :asc).first.valid_from
           else
             valid_to
           end
         }
 
         bitemporal_scope = finder_class.unscoped.ignore_valid_datetime
-            .where("valid_from < ?", valid_to).where("valid_to > ?", valid_from)
+            .where("valid_from < ?", valid_to).where("? < valid_to", valid_from)
             .yield_self { |scope|
               # MEMO: #dup などでコピーした場合、id は存在しないが swapped_id のみ存在するケースがあるので
               # id と swapped_id の両方が存在する場合のみクエリを追加する